Fixed For each loop Issues and Added Java Documentation to Distance
Before there I tried iterating through a hashmap as if it were a list. Adjusted to iterate through a hash map.

Also there was an issue trying to iterate through a default list manager, editted Class Training Example to have an ArrayList identicle to the defualt list manager so it can be used more flexible

Added documentation for class distance as well

Addition to update distances, at the end of updating distances it will automatically normalize